pyproject.toml
[project]
name = "argobeast"
version = "2.1.2"
description = "A Python-based test automation framework for Behave with a clean Page Object architecture and CLI scaffolding."
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Paul Shears", email = "paul.s@argodevops.co.uk" }
]

# Your dependencies for the framework itself (not test requirements)
dependencies = [
    "selenium>=4.0",
    "PyYAML>=6.0",
    "deepmerge>=1.0",
    "python-dotenv>=1.0",
    "behave>=1.2",
]

# Entry point for CLI
[project.scripts]
argobeast = "argo_beast.cli.main:main"

# Setuptools settings
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}  # IMPORTANT for src/ layout

# Include non-Python files inside the package
[tool.setuptools.package-data]
"argo_beast" = ["**/*.yml"]

# Automatically include all packages under src/
[tool.setuptools.packages.find]
where = ["src"]

# Include options dependencies for Sphinx documentation
[project.optional-dependencies]
docs = [
    "sphinx",
    "sphinx-autodoc-typehints",
    "myst_parser" 
]
allure = [
    "allure-behave"
]

[dependency-groups]
dev = [
    "pytest==9.0.3",
    "pylint",
    "build",  # The standard PEP 517 build frontend
    "twine",  # For uploading to PyPI
    "argobeast[allure]", # For pytests to run correctly
]

[tool.pylint.messages_control]
# Disable nitpicky rules that often clash with testing frameworks
disable = [
    "C0114", # missing-module-docstring
    "C0115", # missing-class-docstring
    "C0116", # missing-function-docstring
    "R0903", # too-few-public-methods (common in Page Objects)
    "W0511", # fixme (allow TODOs),
    "W0718",
    "W1203", # logging-fstring-interpolation
]

[tool.pylint.format]
max-line-length = 120  # More generous than the default 100

[tool.pylint.master]
fail-under = 10.0 # Force a perfect score on your filtered rules